requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
appdirs==1.4.4
argcomplete==3.6.3
arrow==1.4.0
asgiref==3.11.0
attr==0.3.1
attrs
azure-core==1.37.0
azure-storage-blob==12.27.1
beautifulsoup4
black==26.1a1
bleach==5.0.1
boto3==1.42.17
botocore==1.42.17
cachetools==6.2.4
certifi
cffi
charset-normalizer
click==8.3.1
colorama
comtypes==1.4.14
contourpy==1.3.3
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
datamodel-code-generator==0.26.1
decorator
defusedxml
distro==1.9.0
Django==5.1.15
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==3.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
dnspython==2.8.0
dotenv==0.9.9
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
dxcam==0.0.5
email-validator==2.3.0
expiringdict==1.2.2
Faker==39.0.0
filelock
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
genson==1.3.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.5.0
google-cloud-logging==3.13.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11
html5lib
httpcore
httpx
idna
ijson==3.4.0.post0
importlib-metadata==8.7.1
inflect==5.6.2
inflection==0.5.1
ipykernel
ipython
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
isort==5.13.2
jedi
Jinja2
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsf==0.11.2
json5
jsonpointer==3.0.0
jsonschema
jsonschema-specifications
jupyterlab
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
label-studio==1.22.0
label-studio-sdk==2.0.16
lark==1.3.1
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
loggerplusplus==1.0.1
loguru==0.7.3
lxml==6.0.2
lxml-html-clean==0.4.3
Markdown==3.10
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
mistune
mss==10.1.0
mypy-extensions==1.1.0
networkx
nltk==3.9.2
notebook
numpy==2.0.1
openai==1.109.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
ordered-set==4.0.2
packaging
pandas==2.3.3
pillow
platformdirs
polars==1.36.1
prometheus-client
proto-plus==1.27.0
protobuf==6.33.2
psutil
psycopg==3.3.2
psycopg-binary==3.3.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyboxen==1.3.0
pydantic==2.12.5
pydantic-core==2.41.5
Pygments
PyJWT==2.10.1
pyparsing==3.3.1
pyRFC3339==2.1.0
python-dateutil
python-dotenv==1.2.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML
pyzmq
redis==5.2.1
regex==2025.11.3
requests
requests-file==3.0.1
requests-mock==1.12.1
rich==14.2.0
rq==2.6.1
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.16.0
scipy==1.16.3
semver==3.0.4
sentry-sdk==2.48.0
six
smart-open==7.5.0
soupsieve
sqlparse==0.5.5
sympy==1.13.1
tldextract==5.3.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing-extensions
typing-inspection==0.4.2
tzdata==2025.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3
user-agents==2.2.0
uuid-utils==0.12.0
webcolors==25.10.0
websocket-client
Werkzeug==3.1.4
wrapt==2.0.1
xmljson==0.2.1
zipp==3.23.0
